+++ lib/pesapal.rb
@@ -1,6 +1,7 @@
 require "rails"
 require "htmlentities"
 require "pesapal/merchant"
+require "pesapal/merchant/details"
 require "pesapal/merchant/post"
 require "pesapal/merchant/status"
 require "pesapal/oauth"