+++ lib/pesapal/merchant/status.rb
@@ -2,12 +2,12 @@ module Pesapal
 
     module Status
         
-        # set parameters required by the QueryPaymentStatus & QueryPaymentStatusByMerchantRef 
+        # set parameters required by the QueryPaymentStatus & QueryPaymentStatusByMerchantRef calls
         def Status.set_parameters(consumer_key, merchant_reference, transaction_tracking_id = nil)
 
-            # parameters required by the PostPesapalDirectOrderV4 call (excludes
+            # parameters required by the QueryPaymentStatus call (excludes
             # oauth_signature parameter as per the instructions here
-            # http://developer.pesapal.com/how-to-integrate/api-reference#PostPesapalDirectOrderV4)
+            # http://developer.pesapal.com/how-to-integrate/api-reference)
             
             timestamp = Time.now.to_i.to_s