+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.2.2
 
 * Automate release date to reduce chances of forgetting to update it
 * Add new functionality: Query Payment Status
+* Add new functionality: Query Payment Details
 
 v0.2.1
 ------